+++ src/models/courses.model.ts
@@ -13,8 +13,8 @@ export default class CoursesModel {
     #createTableIfNotExists () {
         this.client.query(`
             CREATE TABLE IF NOT EXISTS courses (
-                ID SERIAL PRIMARY KEY,
-                USER_ID INT NOT NULL REFERENCES users(id),
+                ID BIGSERIAL PRIMARY KEY,
+                USER_ID BIGINT NOT NULL REFERENCES users(id),
                 GOOGLE_DRIVE_THUMBNAIL_ID VARCHAR(256) NOT NULL,
                 TITLE VARCHAR(128) NOT NULL,
                 DESCRIPTION VARCHAR,
